requirements.txt
torch==1.11.0
torchvision==0.12.0
torchaudio~=0.11.0
librosa
numpy
tqdm
matplotlib
pandas
pillow

yadisk
wandb
gdown